// packages/kaoto-vscode/src/types/IntegrationTreeItemType.ts
/**
 * Copyright 2025 Red Hat, Inc. and/or its affiliates.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *        http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
import { Uri } from 'vscode';

export type IntegrationFileType = 'kamelet' | 'pipe' | 'route';

export type IntegrationFileIconType = 'kamelet' | 'pipe' | 'route' | 'route-child';

export type IntegrationFileIcon = {
	light: Uri;
	dark: Uri;
};

export type IntegrationFileDSL = 'yaml' | 'xml';

export type IntegrationFile = {
	dsl: IntegrationFileDSL;
	type: IntegrationFileType;
	name: string;
	icon: IntegrationFileIcon;
	description: string;
};